bluelotusfoods-api/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
pydantic
pydantic-settings
httpx
structlog
